fix no changes true
