+++ roles/importer/files/importer/model_controllers/fwconfig_import.py
@@ -6,6 +6,7 @@ import fwo_const
 from fwo_log import getFwoLogger
 from model_controllers.import_state_controller import ImportStateController
 from fwo_api_oo import FwoApi
+from fwo_base import ConfigAction, ConfFormat
 from models.fwconfig_normalized import FwConfigNormalized
 from model_controllers.fwconfig_normalized_controller import FwConfigNormalizedController
 from model_controllers.fwconfig_import_object import FwConfigImportObject
@@ -40,13 +41,12 @@ class FwConfigImport(FwConfigImportObject, FwConfigImportRule, FwConfigImportGat
         FwConfigImportGateway.__init__(self, importState, config)
         
     def importConfig(self):
-        # assuming we always get the full config (only inserts) from API
-        previousConfig = self.getPreviousConfig()
-        self.updateDiffs(previousConfig)
-        # calculate differences and write them to the database via API
+        # current implementation restriction: assuming we always get the full config (only inserts) from API
 
-        # rbLinks = RulebaseLinkController.getRulebaseLinks()
-        # TODO: deal with networking later
+        previousConfig = self.getPreviousConfig()
+
+        # calculate differences and write them to the database via API
+        self.updateDiffs(previousConfig)
         return
 
     def updateDiffs(self, previousConfig: FwConfigNormalized):
@@ -188,3 +188,36 @@ class FwConfigImport(FwConfigImportObject, FwConfigImportRule, FwConfigImportGat
     #                      GlobalPolicyUid=None  # TODO: global policy UID
     #                      )
     #         self.NormalizedConfig.gateways.append(gw)
+
+
+    # return previous config or empty config if there is none
+    def getPreviousConfig(self) -> FwConfigNormalized:
+        logger = getFwoLogger(debug_level=self.ImportDetails.DebugLevel)
+        query = "query getLatestConfig($mgmId: Int!) { latest_config(where: {mgm_id: {_eq: $mgmId}}) { config } }"
+        queryVariables = { 'mgmId': self.ImportDetails.MgmDetails.Id }
+        try:
+            queryResult = self.ImportDetails.call(query, queryVariables=queryVariables)
+            if 'errors' in queryResult:
+                logger.exception("fwo_api:import_latest_config - error while deleting last config for mgm id " +
+                                str(self.ImportDetails.MgmDetails.Id) + ": " + str(queryResult['errors']))
+                return 1 # error
+            else:
+                if len(queryResult['data']['latest_config'])>0: # do we have a prev config?
+                    # prevConfigDict = json.loads(queryResult['data']['latest_config'][0]['config'])
+                    prevConfig = FwConfigNormalized.parse_raw(queryResult['data']['latest_config'][0]['config'])
+                else:
+                    prevConfigDict = {
+                        'action': ConfigAction.INSERT,
+                        'network_objects': {},
+                        'service_objects': {},
+                        'users': {},
+                        'zone_objects': {},
+                        'rules': [],
+                        'gateways': [],
+                        'ConfigFormat': ConfFormat.NORMALIZED_LEGACY
+                    }
+                    prevConfig = FwConfigNormalized(**prevConfigDict)
+                return prevConfig
+        except:
+            logger.exception(f"failed to get latest normalized config for mgm id {str(self.ImportDetails.MgmDetails.Id)}: {str(traceback.format_exc())}")
+            raise Exception(f"error while trying to get the previous config")
